fix house

+++ src/main/java/com/itheima/entity/Person.java
@@ -10,5 +10,5 @@ public class Person {
     public int id;
     public int age;
     public boolean sex;
-    public int house;
+    public long house;
 }
